Revert "Backend kafka"
