--- Back-end/microservices-parent/microservices/inventory-service/src/main/java/com/petcove/inventoryservice/event/OrderItemEvent.java
@@ -1,20 +1,0 @@
-package com.petcove.inventoryservice.event;
-
-import lombok.*;
-
-import java.io.Serializable;
-import java.math.BigDecimal;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderItemEvent implements Serializable  {
-    private Long id;
-
-    private String skuCode;
-
-    private Integer quantity;
-
-    private BigDecimal price;
-
-}
