--- Back-end/microservices-parent/microservices/inventory-service/src/main/java/com/petcove/inventoryservice/event/OrderPlacedEvent.java
@@ -1,21 +1,0 @@
-package com.petcove.inventoryservice.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderPlacedEvent implements Serializable {
-
-    private Long orderId;
-    private Long customerId;
-    private BigDecimal totalAmount;
-
-    private List<OrderItemEvent> orderItems;
-}
