--- Back-end/microservices-parent/microservices/inventory-service/src/main/java/com/petcove/inventoryservice/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.petcove.inventoryservice.exception;
-
-public class ProductNotFoundException extends RuntimeException{
-    public ProductNotFoundException() {
-        super("Product not found in the inventory.");
-    }
-}
