+++ Back-end/microservices-parent/microservices/inventory-service/src/main/java/com/petcove/inventoryservice/service/InventoryService.java
@@ -1,7 +1,6 @@
 package com.petcove.inventoryservice.service;
 
 import com.petcove.inventoryservice.dto.InventoryResponse;
-import com.petcove.inventoryservice.model.Inventory;
 import com.petcove.inventoryservice.repository.InventoryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
@@ -9,38 +8,30 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static java.util.stream.Collectors.toList;
 
 @Service
 @RequiredArgsConstructor
-@Slf4j
+//@Slf4j
 public class InventoryService {
 
     private final InventoryRepository inventoryRepository;
     @Transactional(readOnly = true)
     //@SneakyThrows
-    public List<InventoryResponse> isInstock(List<String> skuCode, List<Integer> quantity){
+    public List<InventoryResponse> isInstock(List<String> skuCode){
         /*Simulate time-out exception
         log.info("Wait started");
         Thread.sleep(10000);
         log.info("wait ended");
         */
         //check if the *object git is present inside the optional or not
-        log.info("start isinstock");
-        Map<String, Integer> sku_quan_map = new HashMap<String,Integer>();
-        for(int i=0; i<skuCode.size(); i++){
-            sku_quan_map.put(skuCode.get(i), quantity.get(i));
-        }
-        log.info(sku_quan_map.toString());
         return inventoryRepository.findBySkuCodeIn(skuCode).stream()
                 .map(inventory ->
                     InventoryResponse.builder()
                             .skuCode(inventory.getSkuCode())
-                            .isInStock(inventory.getQuantity() >= sku_quan_map.get(inventory.getSkuCode()))
+                            .isInStock(inventory.getQuantity() > 0)
                             .build()
                 ).toList();
     }
