+++ Back-end/microservices-parent/microservices/notification-service/src/main/java/com/petcove/notificationservice/NotificationServiceApplication.java
@@ -5,8 +5,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.kafka.annotation.KafkaListener;
 
-import java.util.Map;
-
 @SpringBootApplication
 @Slf4j
 public class NotificationServiceApplication {
@@ -16,7 +14,6 @@ public class NotificationServiceApplication {
 
     @KafkaListener(topics = "notificationTopic")
     public void handelNotification(OrderPlacedEvent orderPlacedEvent){
-        //String sq = orderPlacedEvent.getId();
-        //log.info(sq);
+        log.info("Received notification from order - {}", orderPlacedEvent.getOrderNumber());
     }
 }
