+++ Back-end/microservices-parent/microservices/notification-service/src/main/java/com/petcove/notificationservice/OrderPlacedEvent.java
@@ -4,12 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderPlacedEvent {
-    private String id;
+    private String orderNumber;
 }
