+++ Back-end/microservices-parent/microservices/notification-service/src/main/resources/application.properties
@@ -4,8 +4,7 @@ spring.application.name=notification-service
 
 
 spring.kafka.bootstrap-servers=localhost:9092
-spring.kafka.consumer.group-id=placeorderID
+spring.kafka.consumer.group-id=notificationID
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spring.kafka.consumer.value-serializer=org.springframework.kafka.support.serializer.JsonDeserializer
 spring.kafka.consumer.spring.json.type.mapping=event:com.petcove.notificationservice.OrderPlacedEvent
-spring.kafka.consumer.properties.spring.json.trusted.packages=*
