--- Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/event/OrderItemEvent.java
@@ -1,21 +1,0 @@
-package com.petcove.orderservice.event;
-
-import lombok.*;
-
-import java.io.Serializable;
-import java.math.BigDecimal;
-
-
-@Builder
-@Getter
-@Setter
-public class OrderItemEvent implements Serializable {
-    private Long id;
-
-    private String skuCode;
-
-    private Integer quantity;
-
-    private BigDecimal price;
-
-}
