+++ Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/event/OrderPlacedEvent.java
@@ -4,19 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.Map;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class OrderPlacedEvent implements Serializable {
-
-    private Long orderId;
-    private Long customerId;
-    private BigDecimal totalAmount;
-
-    private List<OrderItemEvent> orderItems;
+public class OrderPlacedEvent {
+    private String orderNumber;
 }
