+++ Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/model/Order.java
@@ -1,11 +1,11 @@
 package com.petcove.orderservice.model;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import jakarta.persistence.*;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
@@ -14,28 +14,11 @@ import java.util.List;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(name = "customer_id", nullable = false)
-    private Long customerId;
-
-    @Column(name = "total_amount", nullable = false)
-    private BigDecimal totalAmount;
-
-    //@Column(name = "order_date", nullable = false)
-    //private LocalDateTime orderDate;
-
-    @Column(name = "order_number", nullable = false)
     private String orderNumber;
-
-    @Column(name = "order_status", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private OrderStatus orderStatus;
-
-    @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @OneToMany(cascade = CascadeType.ALL)
     private List<OrderLineItems> orderLineItemsList;
 }
