+++ Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/model/OrderLineItems.java
@@ -1,6 +1,9 @@
 package com.petcove.orderservice.model;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import jakarta.persistence.*;
 import java.math.BigDecimal;
@@ -11,22 +14,11 @@ import java.math.BigDecimal;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class OrderLineItems {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(name = "skuCode", nullable = false)
     private String skuCode;
-
-    @Column(name = "price", nullable = false)
     private BigDecimal price;
-
-    @Column(name = "quantity", nullable = false)
     private Integer quantity;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "order_id")
-    private Order order;
 }
