--- Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/model/OrderStatus.java
@@ -1,8 +1,0 @@
-package com.petcove.orderservice.model;
-
-public enum OrderStatus {
-    Handling,
-    PLACED,
-    PAID,
-    CANCELLED
-}
