+++ Back-end/microservices-parent/microservices/order-service/src/main/java/com/petcove/orderservice/service/OrderService.java
@@ -8,17 +8,63 @@ import com.petcove.orderservice.model.Order;
 import com.petcove.orderservice.model.OrderLineItems;
 import com.petcove.orderservice.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
 
+@Service
+@RequiredArgsConstructor
+public class OrderService {
 
-public interface OrderService {
-    String placeOrder(OrderRequest orderRequest);
+    private final OrderRepository orderRepository;
+    private final WebClient.Builder webClientBuilder;
+    private final KafkaTemplate<String, OrderPlacedEvent> kafkaTemplate;
+    public String placeOrder(OrderRequest orderRequest){
+        Order order = new Order();
+        order.setOrderNumber(UUID.randomUUID().toString());
 
-    OrderLineItems DtoToOrderItems(OrderLineItemsDto orderLineItemsDto);
-    //OrderLineItems mapToDto(OrderLineItemsDto orderLineItemsDto);
+        List<OrderLineItems> orderLineItems = orderRequest.getOrderLineItemsDtoList()
+                .stream()
+                .map(this::mapToDto)
+                .toList();
+
+        order.setOrderLineItemsList(orderLineItems);
+
+        List<String> skuCodes = order.getOrderLineItemsList().stream()
+                .map(OrderLineItems::getSkuCode)
+                .toList();
+
+        // Call Inventory service and place order if the product is in stock; result true means in stock
+        InventoryResponse[] inventoryResponseArr = webClientBuilder.build().get() //synchronous request
+                //build the uri with the query params as skuCodes
+                .uri("http://inventory-service/api/inventory",
+                        uriBuilder -> uriBuilder.queryParam("skuCode",skuCodes).build())
+                .retrieve()
+                .bodyToMono(InventoryResponse[].class) //read the response
+                .block();
+        boolean allProductsInStock = Arrays.stream(inventoryResponseArr)
+                .allMatch(InventoryResponse::isInStock);
+
+        if(allProductsInStock){
+            orderRepository.save(order);
+            // send the order placed event as a msg to the notification topic
+            kafkaTemplate.send("notificationTopic", new OrderPlacedEvent(order.getOrderNumber()));
+            return "Order placed successfully.";
+        }else {
+            throw new IllegalArgumentException("Product is not in stock at the moment.");
+        }
+
+    }
+
+    private OrderLineItems mapToDto(OrderLineItemsDto orderLineItemsDto){
+        OrderLineItems orderLineItems = new OrderLineItems();
+        orderLineItems.setPrice(orderLineItemsDto.getPrice());
+        orderLineItems.setQuantity(orderLineItemsDto.getQuantity());
+        orderLineItems.setSkuCode(orderLineItemsDto.getSkuCode());
+        return orderLineItems;
+    }
 }
