+++ Back-end/microservices-parent/microservices/order-service/src/main/resources/application.properties
@@ -31,4 +31,4 @@ resilience4j.retry.instances.inventory.wait-duration=5s
 spring.kafka.bootstrap-servers=localhost:9092
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
-spring.kafka.producer.properties.spring.json.type.mapping=OrderPlacedEvent:com.petcove.orderservice.event.OrderPlacedEvent
+spring.kafka.producer.spring.json.type.mapping=event:com.petcove.orderservice.event.OrderPlacedEvent
